Correct declaration hoisting
Hoist let, const, var, function within _blocks_, rather than scopes to
ensure scope is maintained when `await` creates a new block (but not a
new scope), but relying on the JS engine to do any further hoisting
according to the semantics of the engine (e.g. ES5 vs 6, strict vs
sloppy). 

This commit requires that consts are actually implemented via `let` as
the declaration may be hoisted above an await-created block. A routine
to check illegal const assignment is in progress